+++ src/Idris/AbsSyntax.hs
@@ -1064,6 +1064,23 @@ allImportDirs = do i <- getIState
                    let optdirs = opt_importdirs (idris_options i)
                    return ("." : reverse optdirs)
 
+addSourceDir :: FilePath -> Idris ()
+addSourceDir fp = do i <- getIState
+                     let opts = idris_options i
+                     let opts' = opts { opt_sourcedirs = nub $ fp : opt_sourcedirs opts  }
+                     putIState $ i { idris_options = opts' }
+
+setSourceDirs :: [FilePath] -> Idris ()
+setSourceDirs fps = do i <- getIState
+                       let opts = idris_options i
+                       let opts' = opts { opt_sourcedirs = nub $ fps  }
+                       putIState $ i { idris_options = opts' }
+
+allSourceDirs :: Idris [FilePath]
+allSourceDirs = do i <- getIState
+                   let optdirs = opt_sourcedirs (idris_options i)
+                   return ("." : reverse optdirs)
+
 colourise :: Idris Bool
 colourise = do i <- getIState
                return $ idris_colourRepl i
@@ -2519,6 +2536,10 @@ getIBCSubDir _ = Nothing
 getImportDir :: Opt -> Maybe String
 getImportDir (ImportDir str) = Just str
 getImportDir _ = Nothing
+
+getSourceDir :: Opt -> Maybe String
+getSourceDir (SourceDir str) = Just str
+getSourceDir _ = Nothing
 
 getPkgDir :: Opt -> Maybe String
 getPkgDir (Pkg str) = Just str